// src/lib/email/signature-asset-types.d.ts
export type SignatureAssetMetadata = {
	altText: string;
	contentId: string;
	createdAt: Date;
	filename: string;
	height: number;
	id: string;
	mailboxId: string;
	previewUrl: string;
	size: number;
	src: string;
	type: string;
	width: number;
};

export type SignatureImageDimensions = {
	height: number;
	width: number;
};